Add files via upload
Library Management System is an application which refers to library systems which are generally small or medium in size. It is used by librarian to manage the library record various transactions like issue of books, return of books, addition of new books, addition of new students etc. Books and student maintenance modules are also included in this system which would keep track of the students using the library and also a detailed description about the books a library contains.. In addition , report module is also included in Library Management System .